\name{Affy.HuEx.1st.v2}
\alias{Affy.HuEx.1st.v2}
\docType{data}
\title{
Affymetrix Human Exon 1.0 ST ChrY probes
}
\description{
Y chromosome probes from the Affymetrix Human Exon 1.0 ST array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCh37.p11.
}

\format{
  A data frame with 3254 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}

\examples{
data(Affy.HuEx.1st.v2)
## maybe str(Affy.HuEx.1st.v2) ; plot(Affy.HuEx.1st.v2) ...
}
\keyword{datasets}
