\name{Affy.HuGene.1st.v1}
\alias{Affy.HuGene.1st.v1}
\docType{data}
\title{
Affymetrix Human Gene 1.0 ST ChrY probes
}
\description{
Y chromosome probes from the Affymetrix Human Gene 1.0 ST array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCh37.p11.
}

\format{
  A data frame with 171 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}

\examples{
data(Affy.HuGene.1st.v1)
## maybe str(Affy.HuGene.1st.v1) ; plot(Affy.HuGene.1st.v1) ...
}
\keyword{datasets}
