\name{Affy.MoEx.1st.v1}
\alias{Affy.MoEx.1st.v1}
\docType{data}
\title{
Affymetrix Mouse Exon 1.0 ST ChrY probes
}
\description{
Y chromosome probes from the Affymetrix Mouse Exon 1.0 ST array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCm38.p1.
}

\format{
  A data frame with 602 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}

\examples{
data(Affy.MoEx.1st.v1)
## maybe str(Affy.MoEx.1st.v1) ; plot(Affy.MoEx.1st.v1) ...
}
\keyword{datasets}
