\name{Affy.U133.plus2}
\alias{Affy.U133.plus2}
\docType{data}
\title{
Affymetrix GeneChip Human Genome U133 Plus 2.0 ChrY probes
}
\description{
Y chromosome probes from the Affymetrix GeneChip Human Genome U133 Plus 2.0 array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCh37.p11.
}

\format{
  A data frame with 6 observations on the following variable.
  \describe{
    \item{\code{V1}}{a factor with levels \code{1555743_s_at} \code{1569609_at} \code{204410_at} \code{207909_x_at} \code{208282_x_at} \code{217162_at}}
  }
}

\examples{
data(Affy.U133.plus2)
## maybe str(Affy.U133.plus2) ; plot(Affy.U133.plus2) ...
}
\keyword{datasets}
