\name{Illumina.Human.WG6.v1}
\alias{Illumina.Human.WG6.v1}
\docType{data}
\title{
Illumina Human WG6 v1 ChrY probes
}
\description{
Y chromosome probes from the Illumina Human WG6 v1 array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCh37.p11.
}

\format{
  A data frame with 62 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}

\examples{
data(Illumina.Human.WG6.v1)
## maybe str(Illumina.Human.WG6.v1) ; plot(Illumina.Human.WG6.v1) ...
}
\keyword{datasets}
