\name{Illumina.Human.WG6.v2}
\alias{Illumina.Human.WG6.v2}
\docType{data}
\title{
Illumina Human WG6 v2 ChrY probes
}
\description{
Y chromosome probes from the Illumina Human WG6 v2 array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCh37.p11.
}

\format{
  A data frame with 100 observations on the following variable.
  \describe{
    \item{\code{V1}}{a factor with levels \code{ILMN_1651456} \code{ILMN_1652263} \code{ILMN_1653471} \code{ILMN_1654284} \code{ILMN_1655210} \code{ILMN_1655513} \code{ILMN_1656021} \code{ILMN_1658123} \code{ILMN_1658315} \code{ILMN_1661523} \code{ILMN_1662052} \code{ILMN_1662151} \code{ILMN_1663406} \code{ILMN_1663598} \code{ILMN_1665736} \code{ILMN_1666140} \code{ILMN_1666417} \code{ILMN_1669640} \code{ILMN_1670821} \code{ILMN_1671067} \code{ILMN_1672935} \code{ILMN_1673390} \code{ILMN_1673417} \code{ILMN_1674666} \code{ILMN_1674678} \code{ILMN_1675852} \code{ILMN_1676960} \code{ILMN_1677061} \code{ILMN_1682809} \code{ILMN_1683214} \code{ILMN_1683596} \code{ILMN_1683872} \code{ILMN_1685690} \code{ILMN_1685986} \code{ILMN_1686911} \code{ILMN_1688833} \code{ILMN_1688861} \code{ILMN_1689843} \code{ILMN_1690753} \code{ILMN_1691375} \code{ILMN_1691598} \code{ILMN_1693537} \code{ILMN_1693622} \code{ILMN_1697937} \code{ILMN_1699687} \code{ILMN_1701242} \code{ILMN_1702929} \code{ILMN_1703983} \code{ILMN_1712370} \code{ILMN_1712799} \code{ILMN_1716969} \code{ILMN_1719688} \code{ILMN_1720983} \code{ILMN_1724931} \code{ILMN_1727258} \code{ILMN_1731521} \code{ILMN_1731709} \code{ILMN_1735767} \code{ILMN_1736175} \code{ILMN_1737543} \code{ILMN_1738703} \code{ILMN_1739587} \code{ILMN_1742460} \code{ILMN_1745914} \code{ILMN_1746858} \code{ILMN_1748182} \code{ILMN_1754528} \code{ILMN_1755537} \code{ILMN_1756506} \code{ILMN_1764541} \code{ILMN_1765734} \code{ILMN_1767355} \code{ILMN_1767590} \code{ILMN_1768489} \code{ILMN_1772163} \code{ILMN_1774902} \code{ILMN_1775851} \code{ILMN_1776195} \code{ILMN_1776719} \code{ILMN_1779475} \code{ILMN_1780365} \code{ILMN_1783033} \code{ILMN_1783142} \code{ILMN_1786311} \code{ILMN_1788506} \code{ILMN_1790515} \code{ILMN_1794407} \code{ILMN_1794651} \code{ILMN_1797207} \code{ILMN_1797563} \code{ILMN_1798314} \code{ILMN_1800243} \code{ILMN_1804958} \code{ILMN_1805519} \code{ILMN_1806313} \code{ILMN_1808923} \code{ILMN_1810319} \code{ILMN_1812328} \code{ILMN_1812760} \code{ILMN_1813093}}
  }
}

\examples{
data(Illumina.Human.WG6.v2)
## maybe str(Illumina.Human.WG6.v2) ; plot(Illumina.Human.WG6.v2) ...
}
\keyword{datasets}
