\name{Illumina.Human.WG6.v3}
\alias{Illumina.Human.WG6.v3}
\docType{data}
\title{
Illumina Human WG6 v3 ChrY probes
}
\description{
Y chromosome probes from the Illumina Human WG6 v3 array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCh37.p11.
}

\format{
  A data frame with 144 observations on the following variable.
  \describe{
    \item{\code{V1}}{a factor with levels \code{ILMN_1651456} \code{ILMN_1653471} \code{ILMN_1655210} \code{ILMN_1656021} \code{ILMN_1658123} \code{ILMN_1658315} \code{ILMN_1661523} \code{ILMN_1662052} \code{ILMN_1662151} \code{ILMN_1663406} \code{ILMN_1663598} \code{ILMN_1665736} \code{ILMN_1666140} \code{ILMN_1666417} \code{ILMN_1669640} \code{ILMN_1670821} \code{ILMN_1671067} \code{ILMN_1672935} \code{ILMN_1673390} \code{ILMN_1673417} \code{ILMN_1674678} \code{ILMN_1675852} \code{ILMN_1676960} \code{ILMN_1677061} \code{ILMN_1678260} \code{ILMN_1683214} \code{ILMN_1683596} \code{ILMN_1683872} \code{ILMN_1685690} \code{ILMN_1688833} \code{ILMN_1688861} \code{ILMN_1689843} \code{ILMN_1690753} \code{ILMN_1691375} \code{ILMN_1693537} \code{ILMN_1697937} \code{ILMN_1703983} \code{ILMN_1716969} \code{ILMN_1719688} \code{ILMN_1720983} \code{ILMN_1724931} \code{ILMN_1727258} \code{ILMN_1731521} \code{ILMN_1736175} \code{ILMN_1737543} \code{ILMN_1738703} \code{ILMN_1739587} \code{ILMN_1742460} \code{ILMN_1745914} \code{ILMN_1748182} \code{ILMN_1750879} \code{ILMN_1754528} \code{ILMN_1755537} \code{ILMN_1756506} \code{ILMN_1762573} \code{ILMN_1764541} \code{ILMN_1765734} \code{ILMN_1767355} \code{ILMN_1767590} \code{ILMN_1768489} \code{ILMN_1772163} \code{ILMN_1774751} \code{ILMN_1775851} \code{ILMN_1776195} \code{ILMN_1779475} \code{ILMN_1780365} \code{ILMN_1783142} \code{ILMN_1788506} \code{ILMN_1790515} \code{ILMN_1794407} \code{ILMN_1796326} \code{ILMN_1800421} \code{ILMN_1804958} \code{ILMN_1805519} \code{ILMN_1806313} \code{ILMN_1808923} \code{ILMN_1810319} \code{ILMN_1812328} \code{ILMN_2048223} \code{ILMN_2048228} \code{ILMN_2052433} \code{ILMN_2052438} \code{ILMN_2055391} \code{ILMN_2055396} \code{ILMN_2056795} \code{ILMN_2063254} \code{ILMN_2066116} \code{ILMN_2068522} \code{ILMN_2077896} \code{ILMN_2080306} \code{ILMN_2080309} \code{ILMN_2087817} \code{ILMN_2089365} \code{ILMN_2090029} \code{ILMN_2090059} \code{ILMN_2091113} \code{ILMN_2094332} \code{ILMN_2097052} \code{ILMN_2098499} \code{ILMN_2099855} \code{ILMN_2099858} \code{ILMN_2117845} \code{ILMN_2143383} \code{ILMN_2145470} \code{ILMN_2145781} \code{ILMN_2149793} \code{ILMN_2156971} \code{ILMN_2163538} \code{ILMN_2163540} \code{ILMN_2164478} \code{ILMN_2164482} \code{ILMN_2170157} \code{ILMN_2174266} \code{ILMN_2174275} \code{ILMN_2175124} \code{ILMN_2186846} \code{ILMN_2191331} \code{ILMN_2194108} \code{ILMN_2194115} \code{ILMN_2197608} \code{ILMN_2197750} \code{ILMN_2202368} \code{ILMN_2202370} \code{ILMN_2204360} \code{ILMN_2205022} \code{ILMN_2210199} \code{ILMN_2211956} \code{ILMN_2217056} \code{ILMN_2224230} \code{ILMN_2224231} \code{ILMN_2224454} \code{ILMN_2228976} \code{ILMN_2254722} \code{ILMN_2264634} \code{ILMN_2288142} \code{ILMN_2288285} \code{ILMN_2288290} \code{ILMN_2291534} \code{ILMN_2291538} \code{ILMN_2317524} \code{ILMN_2336306} \code{ILMN_2363185} \code{ILMN_2401946} \code{ILMN_2404951}}
  }
}

\examples{
data(Illumina.Human.WG6.v3)
## maybe str(Illumina.Human.WG6.v3) ; plot(Illumina.Human.WG6.v3) ...
}
\keyword{datasets}
