\name{Illumina.Mouse.WG6.v2}
\alias{Illumina.Mouse.WG6.v2}
\docType{data}
\title{
Illumina Mouse WG6 v2 ChrY probes
}
\description{
Y chromosome probes from the Illumina Mouse WG6 v2 array. The probes in this dataset mapped to the Y chromosome of Ensembl reference genome GRCm38.p1.
}

\format{
  A data frame with 49 observations on the following variable.
  \describe{
    \item{\code{V1}}{a factor with levels \code{ILMN_1213657} \code{ILMN_1215338} \code{ILMN_1215914} \code{ILMN_1217982} \code{ILMN_1224210} \code{ILMN_1224484} \code{ILMN_1225693} \code{ILMN_1227623} \code{ILMN_1228621} \code{ILMN_1228766} \code{ILMN_1229648} \code{ILMN_1230735} \code{ILMN_1234015} \code{ILMN_1235995} \code{ILMN_1239158} \code{ILMN_1241001} \code{ILMN_1241604} \code{ILMN_1242126} \code{ILMN_1243580} \code{ILMN_1244760} \code{ILMN_1246408} \code{ILMN_1246528} \code{ILMN_1247672} \code{ILMN_1250448} \code{ILMN_1252709} \code{ILMN_1254824} \code{ILMN_1255639} \code{ILMN_1257342} \code{ILMN_1259575} \code{ILMN_2467512} \code{ILMN_2487781} \code{ILMN_2540740} \code{ILMN_2540756} \code{ILMN_2540775} \code{ILMN_2540789} \code{ILMN_2540844} \code{ILMN_2563323} \code{ILMN_2588056} \code{ILMN_2668493} \code{ILMN_2758581} \code{ILMN_2788502} \code{ILMN_2819553} \code{ILMN_2866712} \code{ILMN_2879232} \code{ILMN_2899599} \code{ILMN_2899607} \code{ILMN_2907351} \code{ILMN_2922284} \code{ILMN_2998658}}
  }
}

\examples{
data(Illumina.Mouse.WG6.v2)
## maybe str(Illumina.Mouse.WG6.v2) ; plot(Illumina.Mouse.WG6.v2) ...
}
\keyword{datasets}
