\name{MASSI.test.dataset}
\alias{MASSI.test.dataset}
\docType{data}
\title{
MASSI.test.dataset
}
\description{
This data frame is a sample/probe expression matrix, with samples names as column names and probe id's as row names. This is the format required to begin any MASSI analysis.
}
\usage{data(MASSI.test.dataset)}
\format{
  A data frame with 1026 observations on the following 60 variables.
  \describe{
    \item{\code{S1}}{a numeric vector}
    \item{\code{S2}}{a numeric vector}
    \item{\code{S3}}{a numeric vector}
    \item{\code{S4}}{a numeric vector}
    \item{\code{S5}}{a numeric vector}
    \item{\code{S6}}{a numeric vector}
    \item{\code{S7}}{a numeric vector}
    \item{\code{S8}}{a numeric vector}
    \item{\code{S9}}{a numeric vector}
    \item{\code{S10}}{a numeric vector}
    \item{\code{S11}}{a numeric vector}
    \item{\code{S12}}{a numeric vector}
    \item{\code{S13}}{a numeric vector}
    \item{\code{S14}}{a numeric vector}
    \item{\code{S15}}{a numeric vector}
    \item{\code{S16}}{a numeric vector}
    \item{\code{S17}}{a numeric vector}
    \item{\code{S18}}{a numeric vector}
    \item{\code{S19}}{a numeric vector}
    \item{\code{S20}}{a numeric vector}
    \item{\code{S21}}{a numeric vector}
    \item{\code{S22}}{a numeric vector}
    \item{\code{S23}}{a numeric vector}
    \item{\code{S24}}{a numeric vector}
    \item{\code{S25}}{a numeric vector}
    \item{\code{S26}}{a numeric vector}
    \item{\code{S27}}{a numeric vector}
    \item{\code{S28}}{a numeric vector}
    \item{\code{S29}}{a numeric vector}
    \item{\code{S30}}{a numeric vector}
    \item{\code{S31}}{a numeric vector}
    \item{\code{S32}}{a numeric vector}
    \item{\code{S33}}{a numeric vector}
    \item{\code{S34}}{a numeric vector}
    \item{\code{S35}}{a numeric vector}
    \item{\code{S36}}{a numeric vector}
    \item{\code{S37}}{a numeric vector}
    \item{\code{S38}}{a numeric vector}
    \item{\code{S39}}{a numeric vector}
    \item{\code{S40}}{a numeric vector}
    \item{\code{S41}}{a numeric vector}
    \item{\code{S42}}{a numeric vector}
    \item{\code{S43}}{a numeric vector}
    \item{\code{S44}}{a numeric vector}
    \item{\code{S45}}{a numeric vector}
    \item{\code{S46}}{a numeric vector}
    \item{\code{S47}}{a numeric vector}
    \item{\code{S48}}{a numeric vector}
    \item{\code{S49}}{a numeric vector}
    \item{\code{S50}}{a numeric vector}
    \item{\code{S51}}{a numeric vector}
    \item{\code{S52}}{a numeric vector}
    \item{\code{S53}}{a numeric vector}
    \item{\code{S54}}{a numeric vector}
    \item{\code{S55}}{a numeric vector}
    \item{\code{S56}}{a numeric vector}
    \item{\code{S57}}{a numeric vector}
    \item{\code{S58}}{a numeric vector}
    \item{\code{S59}}{a numeric vector}
    \item{\code{S60}}{a numeric vector}
  }
}
\details{
This data is from an Illumina human-6 v2.0 expression beadchip.
}
\source{
Data modified and obtained from GSE25906
<http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE25906>
}
\references{
Tsai S, Hardison NE, James AH, Motsinger-Reif AA et al. Transcriptional profiling of human placentas from pregnancies complicated by preeclampsia reveals disregulation of sialic acid acetylesterase and immune signalling pathways. Placenta 2011 Feb;32(2):175-82. PMID: 21183218
}
\examples{
data(MASSI.test.dataset)
## maybe str(MASSI.test.dataset) ; plot(MASSI.test.dataset) ...
}
\keyword{datasets}
