\name{MASSI.test.probes}
\alias{MASSI.test.probes}
\docType{data}
\title{
MASSI.test.probes
}
\description{
A list of Y chromosome probes corresponding to the MASSI.test.dataset
}
\usage{data(MASSI.test.probes)}
\format{
  A data frame with 56 observations on the following 0 variables.
  }

\details{
These Y chromosome probe Id's correspond to the Illumina human-6 v2.0 expression beadchip.
}

\examples{
data(MASSI.test.probes)
## maybe str(MASSI.test.probes) ; plot(MASSI.test.probes) ...
}
\keyword{datasets}
